ML/requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
numpy==1.26.4
torch>=2.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy==2.0.35
pydantic==2.9.2
httpx==0.27.2